+++ src/pages/Contact.js
@@ -1,11 +1,22 @@
-import Contact from '../components/Contact';
+import styled from 'styled-components';
+import Collection from '../components/contact/Collection';
+import FormSection from '../components/contact/FormSection';
+import { MainWrapper } from '../styles';
 
-const Contact_page = () => {
+const Contact = () => {
    return (
-      <>
-         <Contact />
-      </>
+      <ContactWarpper>
+         <FormSection />
+         <Collection />
+      </ContactWarpper>
    );
 };
 
-export default Contact_page;
+const ContactWarpper = styled(MainWrapper)`
+   display: flex;
+   justify-content: center;
+   align-items: center;
+   color: white;
+`;
+
+export default Contact;
